requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0
python-multipart>=0.0.9
email-validator>=2.1.0
razorpay>=1.4.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0
